LayoutTests/fast/frames/iframe-detached-window-still-writable-expected.txt

Verify that detached windows are still writable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframeContentWindow.foo is 2
PASS successfullyParsed is true

TEST COMPLETE
